On branch poi-page-optimization  Your branch is behind 'origin/poi-page-optimization' by 43 commits, and can be fast-forwarded.    (use "git pull" to update your local branch)
 Changes to be committed:
	modified:   luna-web/src/main/webapp/manage_poi.jsp
	modified:   luna-web/src/main/webapp/scripts/manage_poi_edit.js
	modified:   luna-web/src/main/webapp/styles/common.css

+++ luna-web/src/main/webapp/scripts/manage_poi_edit.js
@@ -135,7 +135,7 @@ function newBlankPoiReadOnly(_id) {
 };
 
 //“删除”按钮，删除POI数据
-function delPOI(obj, _id){
+function delPOI(obj, _id, poiName){
 	$.ajax({
         url: Util.strFormat( Inter.getApiUrl().poiCheckDelete.url,[_id]),
         type: 'GET',
@@ -152,6 +152,11 @@ function delPOI(obj, _id){
                     popWindow($popwindow);
                     //弹窗中的确定按钮
                     $("#btn-delete").unbind().click(function(){
+						var reason = $('.delete-reason').val();
+						if(! reason){
+
+							return;
+						}
                         $.ajax({
                 	        url: Util.strFormat( Inter.getApiUrl().poiDelete.url,[_id]),
                 	        type: 'DELETE',
